Refactors namespace
