+++ backend/GameWave.Api/Contracts/ICompanyService.cs
@@ -1,6 +1,6 @@
 ﻿using GameWave.ObjectModel;
 
-namespace GameWave.Api.Services
+namespace GameWave.API.Contracts
 {
     public interface ICompanyService : IService<Company>, IServiceAsync<Company>
     { }
