+++ backend/GameWave.Api/Contracts/IGameService.cs
@@ -1,6 +1,6 @@
 ﻿using GameWave.ObjectModel;
 
-namespace GameWave.Api.Services
+namespace GameWave.API.Contracts
 {
     public interface IGameService : IService<Game>, IServiceAsync<Game>
     { }
