+++ backend/GameWave.Api/Contracts/IGenreService.cs
@@ -1,6 +1,6 @@
 ﻿using GameWave.ObjectModel;
 
-namespace GameWave.Api.Services
+namespace GameWave.API.Contracts
 {
     public interface IGenreService : IService<Genre>, IServiceAsync<Genre>
     { }
