+++ backend/GameWave.Api/Contracts/IPlatformService.cs
@@ -1,6 +1,6 @@
 ﻿using GameWave.ObjectModel;
 
-namespace GameWave.Api.Services
+namespace GameWave.API.Contracts
 {
     public interface IPlatformService : IService<Platform>, IServiceAsync<Platform>
     { }
