+++ backend/GameWave.Api/Controllers/CompanyController.cs
@@ -8,11 +8,11 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 
-using GameWave.Api.Extensions;
-using GameWave.Api.Services;
+using GameWave.API.Contracts;
+using GameWave.API.Extensions;
 using GameWave.ObjectModel;
 
-namespace GameWave.Api.Controllers
+namespace GameWave.API.Controllers
 {
     [Route("api/v1/[controller]")]
     [ApiController]
