+++ backend/GameWave.Api/Extensions/AssemblyExtensions.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 
-namespace GameWave.Api.Extensions
+namespace GameWave.API.Extensions
 {
     internal static class AssemblyExtensions
     {
