+++ backend/GameWave.Api/Extensions/IQueryableExtensions.cs
@@ -1,7 +1,7 @@
 ﻿using System.Linq;
 using System.Linq.Expressions;
 
-namespace GameWave.Api.Extensions
+namespace GameWave.API.Extensions
 {
     public static class IQueryableExtensions
     {
