+++ backend/GameWave.Api/Program.cs
@@ -1,11 +1,9 @@
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
 using Serilog;
 
-namespace GameWave.Api
+namespace GameWave.API
 {
     public class Program
     {
