--- backend/GameWave.Api/Services/IAgeRatingContentDescriptorService.cs
@@ -1,7 +1,0 @@
-﻿using GameWave.ObjectModel;
-
-namespace GameWave.Api.Services
-{
-    public interface IAgeRatingContentDescriptorService : IService<AgeRatingContentDescriptor>, IServiceAsync<AgeRatingContentDescriptor>
-    { }
-}
