--- backend/GameWave.Api/Services/IAgeRatingService.cs
@@ -1,7 +1,0 @@
-﻿using GameWave.ObjectModel;
-
-namespace GameWave.Api.Services
-{
-    public interface IAgeRatingService : IService<AgeRating>, IServiceAsync<AgeRating>
-    { }
-}
